Remove `minOccurs="1"` constraint for `auc:AllResourceTotal` element
This resolves a bug in XML generation using Soap4R-autogenerated code.

Specifically, the "minOccurs" constraint causes the XML generator to always write a `auc:AllResourceTotals` element, even if it does not contain any `auc:AllResourceTotal` child elements, resulting in a schema validation error.

This is/was the only occurrence of `minOccurs="1"` in the XSD file.